feat: cameraTilesService with shared pmtiles protocol + shared cone geometry
