+++ src/components/map/layers/CameraMarkerLayers.tsx
@@ -2,61 +2,8 @@ import { useMemo, useCallback, useEffect, useRef } from 'react';
 import { Source, Layer, useMap } from 'react-map-gl/maplibre';
 import maplibregl from 'maplibre-gl';
 import { useMapStore } from '../../../store';
-import { DIRECTIONAL_ZONE, CAMERA_DETECTION, ZONE_SAFETY_MULTIPLIERS } from '../../../services/routingConfig';
 import type { ALPRCamera } from '../../../types';
-
-// Helper to create a direction cone polygon from a point and direction
-// Uses the same parameters as the routing algorithm for consistency
-function createDirectionCone(
-  lon: number,
-  lat: number,
-  direction: number,
-  // Use routing config values so visualization matches what routing avoids
-  lengthMeters: number = CAMERA_DETECTION.routeBufferMeters * ZONE_SAFETY_MULTIPLIERS.block * 0.75,
-  spreadDegrees: number = DIRECTIONAL_ZONE.cameraFovDegrees
-): GeoJSON.Feature<GeoJSON.Polygon> {
-  const earthRadius = 6371000; // meters
-  const latRad = (lat * Math.PI) / 180;
-
-  // Convert meters to degrees (approximate)
-  const lengthDeg = (lengthMeters / earthRadius) * (180 / Math.PI);
-
-  // Calculate the three points of the cone
-  const points: [number, number][] = [[lon, lat]]; // Start at camera
-
-  // Left edge of cone
-  const leftAngle = ((direction - spreadDegrees / 2) * Math.PI) / 180;
-  const leftLon = lon + lengthDeg * Math.sin(leftAngle) / Math.cos(latRad);
-  const leftLat = lat + lengthDeg * Math.cos(leftAngle);
-  points.push([leftLon, leftLat]);
-
-  // Create arc for the front of the cone
-  const steps = 8;
-  for (let i = 1; i < steps; i++) {
-    const angle = ((direction - spreadDegrees / 2 + (spreadDegrees * i) / steps) * Math.PI) / 180;
-    const arcLon = lon + lengthDeg * Math.sin(angle) / Math.cos(latRad);
-    const arcLat = lat + lengthDeg * Math.cos(angle);
-    points.push([arcLon, arcLat]);
-  }
-
-  // Right edge of cone
-  const rightAngle = ((direction + spreadDegrees / 2) * Math.PI) / 180;
-  const rightLon = lon + lengthDeg * Math.sin(rightAngle) / Math.cos(latRad);
-  const rightLat = lat + lengthDeg * Math.cos(rightAngle);
-  points.push([rightLon, rightLat]);
-
-  // Close the polygon
-  points.push([lon, lat]);
-
-  return {
-    type: 'Feature',
-    properties: {},
-    geometry: {
-      type: 'Polygon',
-      coordinates: [points],
-    },
-  };
-}
+import { createDirectionCone, parseDirections } from './cameraGeometry';
 
 // Convert cameras to GeoJSON - optimized with pre-allocated array
 function camerasToGeoJSON(cameras: ALPRCamera[]): GeoJSON.FeatureCollection {
@@ -334,9 +281,7 @@ export function CameraMarkerLayers({ cameras, visible, clustered, crossfadeZoom 
 
     const features: GeoJSON.Feature[] = [];
     for (const camera of camerasWithDirection) {
-      const bearings = camera.directions && camera.directions.length > 1
-        ? camera.directions
-        : [camera.direction!];
+      const bearings = parseDirections(camera.direction, camera.directions);
       const ts = camera.osmTimestamp ? new Date(camera.osmTimestamp).getTime() : 0;
       for (const bearing of bearings) {
         const cone = createDirectionCone(camera.lon, camera.lat, bearing);
